data-pipeline/cloud-function-validate/requirements.txt
google-cloud-storage
google-cloud-pubsub
jsonschema
pandas
google-cloud-bigquery

